Implement device class hierachy
In order to keep our codebase sane and easy to debug, we want to
encapsulate all of our sensors/servos/switches/etc into classes, so we
can keep the implementation details separate from all the management
logic. This commit introduces the following classes (located in
devices.h):

Param:
A limit-constrained value object, our basic unit in the
hierarchy. Templated so we can use floats and ints, but don't expect
bools or more complex types to work right. There is a basic unit test
included to check the constraint logic. Param uses std::shared_ptr under
the covers so we can copy them and pass around references safely. This
also allows us to set up chains of Params, where one Param becomes a
limit for another.

Device:
The basic part of most components, which is just a pin and a given
source of input. Includes some extras like an enable() option and a
built-in Timer.

Transducer:
A two-way converter for a common value (signal converter). We expec to
have many sensors that read input in a format that's not very easy to
work with, like ADC. We would rather run the rest of our logic in
something different, like psi. Transducer and it's children are
templated on both the input and output type, so we can convert back and
forth between floating-point and integer data sources.

Sensor:
A type of Transducer which reads from an external input (something on
the car). This is the base class for anything that measures real-world
data or electrical signals. Provides filtering capability.

AnalogSensor:
A Sensor which reads from an analog source. This will normally be an
analogRead() from some pin, but this class can be simulated with the
potentiometer or the touchscreen.

PressureSensor:
An AnalogSensor which measures the current brake pressure, converting
from native ADC to psi for downstream systems. This is our first actual
concrete instance in the hierarchy, so hopefully it will make a good
example.

There are also some other classes which have been sketched out but not
worked through yet, including:
- ServoPWM/JagMotor
- Toggle/InToggle/OutToggle
- Controller/HotrRC/Joystick

At this point, PressureSensor has replaced the existing code in the main
codebase, everything compiles and runs appropriately (changing modes
correctly, not crashing, etc). The math needs to be checked still, and
there are lots of open questions on how to improve on these initial
classes (for instance, we implemented Transducer with two different
params, but bobby suggested using just one and converting on the fly,
and i'm starting to think it's a good idea...). However, this should get
us up and running; if we can change a bunch of our sensors and devices
over to this model we'll end up with a much cleaner codebase. All in the
name of actually being able to debug stuff out in the dust!