Having these little servos hooked to the pwm motor outputs on my dev board makes it easy to spot problems with the outputs. I noticed with every braking release the motor behavior was quite reversed. And sure enough there was an eggregious bug in one of my map statements which likely is responsible for our broken actuator.  Needless to say I fixed that, and now the brake behavior makes a lot more sense.
Also tweaked a bunch of other brake related parameters in the hopes of making things go better once we eventually have a working motor again and try this out on the actual car.

I couldn't help but re-compactify several enum statements
{
	that
	had
	been
	formatted
	to
	obnoxiously
	fill
	as
	much
	of
	my
	screen
	as
	possible
};

In our transition to using objects for the devices, somehow we lost all sub-whole number precision in the speedo and tach, affecting both the values displayed as well as the editability using the touch tuner. I tried to debug this, but ran into a wall because I don't understand a lot of the code in devices.h . Anders you should teach me about how it works, probably a great lesson would be if we debugged this together over a zoom call?