In progress: switch from adafruit neopixel library to fastled, to massively reduce bus data sent to neopixel strip. This doesn't compile at the moment.
Completed transition from Adafruit_Neopixel library to FastLED library, and it compiles (with one inconsequential compiler warning), but on boot the code hangs when neopixels are initialized. Could this be due to its use of RMT unit which is already being used to read PWM pulses?  That'd be lame because the adafruit library also used the RMT unit.  Anyway, needs debugging.