Refactored handlers for ignition, syspower, and especially starter, with support for remote start.  However, somehow reorganizing the hotrc variable arrays broke the controller value math, seg faulting now
Got it working again but had to stop centering controller data that's within the deadzone, and also disable the spike filter to make it work. Needs to be looked at more. Might be best to undo all the hotrc changes in the last couple commits and keep the rest of the changes.

Fixed deadband issue and spike filter and crashing. Side effect of this fix is a significant improvement of  ability to do fine near-zero adjustments on both axes.

simplified loop/conditional logic of hotrc handler leveraging array indices, very clean

Fixed radio lost detection which had broken, and some more cleanup