dodn't save a file
back stepped to a less overwhelming approach to new motor classes.
In progress

completely overhauling include dependencies, while in the middle of writing new motor drivers. This has all the makings of a total shit show, likely doomed to failure, so saving in a new branch.

checkin blah

Holy miracle, I got it to compile. I hopnestly can't believe it. I'm scared to try it.

Tested and mostly it works great. The only known issue is the gas servo is locked at max, even though the code is sending the right us to it as I pull the trigger in stall mode. Debugging ...